use advancedRunPp() instead of PpToolInstance to avoid eslint rule-of-hooks violations
